Update patch set 1
Patch Set 1:

Aaron's test is good enough for me.  Only thing holding this back is the blob side of things.

If someone pullreqs the change against DC I'll be able to merge this from web UIs without digging around in the blob repos myself.  (leaving in 5 minutes for the day...)

Patch-set: 1